# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-Commerce",
    "description": "E-Commerce",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/cart/add": {
      "post": {
        "tags": [
          "E-Commerce Cart's"
        ],
        "description": "Carts in E-Commerce",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Cart's we want to Add",
            "schema": {
              "$ref": "#/definitions/addCart"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required":true
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {}     
      }
      },
      "/cart/view": {
        "get": {
          "tags": [
            "E-Commerce Cart's"
          ],
          "description": "List Carts in E-Commerce",
          "parameters": [

            {
              "name": "authorization",
              "in": "header",
              "required":true
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {}     
        }
        }    
},
  "definitions": { 
...
    "addCart": {
      "required": ["product_id","quantity","user_id"],
      "properties": {
        "product_id": {
          "type": "array",
          "items": {"type": "string"}
        },
        "quantity": {
          "type": "string"
        },
        "user_id": {
          "type": "array",
          "items": {"type": "string"}
        }
      }
    }
  }
}